// src/interfaces/routeApi/response/ICategory.ts
// same for IBrand

// Single category/item in the "data" array
export interface ICategory {
  _id: string;
  name: string;
  slug: string;
  image: string;
  createdAt?: string; // ISO date string
  updatedAt?: string; // ISO date string
}

// // Single category/item in the "data" array
// export interface IBrand {
//   _id: string;
//   name: string;
//   slug: string;
//   image: string;
//   createdAt: string; // ISO date string
//   updatedAt: string; // ISO date string
// }